pyproject.toml
[project]
name = "release-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "dotenv>=0.9.9",
    "pygithub>=2.6.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
